Replace lalrpop with a recursive descent parser
This was long overdue and is pretty much required for improvements
in the error handling realm to become a reality.

The new parser is roughly based on the final lalrpop grammar with
some changes to make it more anemable to LL parsing.

It passes the test suite, and additionally compiles `maplebricks'
and `maplevga' correctly.